idek man

+++ src/main.gd
@@ -13,6 +13,8 @@ func _ready() -> void:
 	current_building = null
 	randomize_building()
 	
+	$PurchaseDisplay/ModeButton.pressed.connect(toggle_delete_mode)
+	
 	pass # Replace with function body.
 
 
@@ -63,7 +65,7 @@ func _process(delta: float) -> void:
 			
 			if not has_neighbour:
 				for cell in building_cells:
-					#cell.show_x = true
+					cell.show_darken = true
 					pass
 			
 			if valid_position and has_neighbour:
@@ -72,30 +74,28 @@ func _process(delta: float) -> void:
 						place_building(building_cells)
 						randomize_building()
 						grid.validate_clusters()
+					else:
+						$PurchaseDisplay/BuildingCost/ShakeTimer.start() 
+						pass
 					
 				Input.set_default_cursor_shape(Input.CURSOR_POINTING_HAND)
 		else:
 			var hovered_cell: Cell = grid.contents[mouse_pos.y][mouse_pos.x]
 			
-			#print(Building.BuildingType.find_key(hovered_cell.type))
-			var dijk_cell = grid.dijkstra(hovered_cell, grid.root_cell)
-			
-			#print(dijk_cell)
-			
-			while(dijk_cell != null):
-				#print("hi")
-				dijk_cell.highlight(Building.BuildingType.Invalid)
-				dijk_cell = dijk_cell.dijk_prev
-			
 			var hovered_cluster: Cluster = hovered_cell.cluster
 			
 			if hovered_cluster != null:
 				for cell: Cell in hovered_cluster.cells:
+					#cell.highlight(Building.BuildingType.Invalid)
 					cell.show_x = true
 				
 				if Input.is_action_just_pressed("left_click"):
-					hovered_cluster.delete()
-					grid.validate_clusters()
+					if Global.resources >= Global.delete_cluster_cost:
+						hovered_cluster.delete()
+						grid.validate_clusters()
+						Global.increment_deleted_clusters()
+					else:
+						$PurchaseDisplay/ModeButton/Label/ShakeTimer.start()
 			
 			pass
